Added sortableJS

+++ src/app/router/editor/editor.component.ts
@@ -1,9 +1,12 @@
-import { ChangeDetectorRef, Component, HostListener, OnInit, QueryList, ViewChildren } from '@angular/core';
+import { ChangeDetectorRef, Component, ElementRef, HostListener, OnInit, QueryList, ViewChild, ViewChildren } from '@angular/core';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { NzUploadChangeParam } from 'ng-zorro-antd/upload';
 import { PageviewComponent } from '../../components/pageview/pageview.component';
 import { EditorService } from '../../services/editor.service';
 import { ALLOWED_FILE_TYPES, ALLOWED_IMAGE_TYPES } from '../../utils/file-utils';
+import Sortable, { MultiDrag } from 'sortablejs';
+import { PDFPageProxy } from 'pdfjs-dist';
+Sortable.mount(new MultiDrag);
 
 @Component({
   selector: 'pdf-editor',
@@ -14,9 +17,13 @@ export class EditorComponent implements OnInit {
 
   supportedFiles = ALLOWED_FILE_TYPES.join(',');
   isLoading = false;
+  chosenPages = [] as HTMLElement[];
+  chosenPageIndices = [] as number[];
 
   @ViewChildren('fullpage') fullpageViewComponents!: QueryList<PageviewComponent>;
+  @ViewChild('previewList') previewListElementRef!: ElementRef;
   loadingMessageId: string | undefined;
+  sortable!: Sortable;
 
   constructor(
     public editorService: EditorService,
@@ -34,8 +41,10 @@ export class EditorComponent implements OnInit {
         if (event.altKey) {
           this.moveCurrentPageDown();
         } else {
-          this.editorService.selectNextPage();
-          this.moveToPage(this.editorService.currentPageIdx);
+          const { currentPageIdx } = this.editorService;
+          const newPageIdx = currentPageIdx + 1;
+          this.selectPage(newPageIdx);
+          this.moveToPage(newPageIdx);
         }
         break;
       case 'k':
@@ -43,8 +52,10 @@ export class EditorComponent implements OnInit {
         if (event.altKey) {
           this.moveCurrentPageUp();
         } else {
-          this.editorService.selectPreviousPage();
-          this.moveToPage(this.editorService.currentPageIdx);
+          const { currentPageIdx } = this.editorService;
+          const newPageIdx = currentPageIdx - 1;
+          this.selectPage(newPageIdx);
+          this.moveToPage(newPageIdx);
         }
         break;
       case 'l':
@@ -79,6 +90,7 @@ export class EditorComponent implements OnInit {
     this.isLoading = false;
     this.messageService.remove(this.loadingMessageId);
     this.messageService.success(`Successfully added ${addedPages} page${addedPages > 1 ? 's' : ''}`, { nzDuration: 1800 });
+    this.initSortable();
   }
 
   handleChange({ file, fileList }: NzUploadChangeParam): void {
@@ -102,9 +114,15 @@ export class EditorComponent implements OnInit {
   }
 
   deleteCurrentPage() {
+    const idx = this.editorService.currentPageIdx;
+    if (idx < 0 || idx >= this.editorService.pages.length) return;
     this.editorService.deleteCurrentPage();
+    this.cdr.detectChanges();
+    this.chosenPageIndices = [];
+    this.chosenPages = [];
     if (this.editorService.pages.length > 0) {
       this.moveToPage(this.editorService.currentPageIdx);
+      this.selectPage(this.editorService.currentPageIdx);
     }
   }
 
@@ -122,19 +140,73 @@ export class EditorComponent implements OnInit {
 
   selectAndMoveToPage(idx: number) {
     if (idx < 0 || idx >= this.editorService.pages.length) return;
-    this.editorService.selectPage(idx);
+    this.cdr.detectChanges();
+    this.selectPage(idx);
     this.moveToPage(idx);
   }
 
+  selectPage(idx: number) {
+    if (idx < 0 || idx >= this.editorService.pages.length) return;
+    this.editorService.selectPage(idx);
+    this.chosenPages.forEach(page => Sortable.utils.deselect(page));
+    const chosenPage = this.previewListElementRef.nativeElement.children[idx];
+    this.chosenPages = [chosenPage];
+    Sortable.utils.select(chosenPage);
+  }
+
   moveToPage(idx: number) {
+    if (idx < 0 || idx >= this.editorService.pages.length) return;
     this.cdr.detectChanges();
     const nativeElement = this.fullpageViewComponents.get(idx)?.elementRef.nativeElement as HTMLElement;
     nativeElement.scrollIntoView();
   }
 
   async ngOnInit(): Promise<void> {
-    // await this.editorService.setSampleFile();
-    // this.moveToPage(0);
+    await this.editorService.setSampleFile();
+    this.moveToPage(0);
+    this.initSortable();
   }
 
+  onBlur() {
+    console.log("asdfasdf");
+  }
+
+  private initSortable() {
+    this.cdr.detectChanges();
+    if (this.sortable) this.sortable.destroy();
+    this.sortable = Sortable.create(this.previewListElementRef.nativeElement, {
+      multiDrag: true,
+      // multiDragKey: 'ctrl' as any,
+      selectedClass: 'selected',
+      chosenClass: 'chosen',
+      delayOnTouchOnly: true,
+      // delay: 300,
+      animation: 199,
+      onSelect: (evt) => {
+        this.chosenPages = evt.items;
+        this.chosenPageIndices = evt.newIndicies.map((o) => o.index);
+        if (this.chosenPageIndices.length === 0) return;
+        this.editorService.currentPageIdx = this.chosenPageIndices.at(-1)!;
+        this.moveToPage(this.editorService.currentPageIdx);
+      },
+      onDeselect: (evt) => {
+        this.chosenPages = evt.items;
+        this.chosenPageIndices = evt.newIndicies.map((o) => o.index);
+        if (this.chosenPageIndices.length === 0) return;
+        this.editorService.currentPageIdx = this.chosenPageIndices.at(-1)!;
+        this.moveToPage(this.editorService.currentPageIdx);
+      },
+      onUpdate: (evt: Sortable.SortableEvent) => {
+        const idcs = this.sortable.toArray().map(id => parseInt(id));
+        const pages = [] as PDFPageProxy[];
+        idcs.forEach(idx => pages.push(this.editorService.pages[idx]));
+        this.editorService.pages = [...pages];
+        console.log(evt);
+        evt.items.forEach((item) => {
+          Sortable.utils.deselect(item);
+        });
+        this.cdr.detectChanges();
+      }
+    });
+  }
 }
